+++ packages/terra-collapsible-menu-view/CHANGELOG.md
@@ -4,6 +4,9 @@ ChangeLog
 Unreleased
 ----------
 ### Changed
+* Replaced Terra-Form with Terra-Form-Checkbox in CollapsibleMenuViewToggle
+
+### Changed
 * Update tests for terra-toolkit v5 and terra-dev-site v5 changes
 
 6.1.0 - (May 21, 2019)